chore: 전체적인 패키지 구조 변경
