--- src/main/java/Player.java
@@ -1,6 +1,0 @@
-public class Player extends Gamer {
-
-    public Player(String name, PlayerCards cards) {
-        super(name, cards);
-    }
-}
