--- src/main/java/PlayerCards.java
@@ -1,7 +1,0 @@
-import java.util.List;
-
-public class PlayerCards extends GamerCards {
-    public PlayerCards(List<Card> cards) {
-        super(cards);
-    }
-}
