--- src/main/java/WinState.java
@@ -1,6 +1,0 @@
-public enum WinState {
-
-    WIN,
-    DRAW,
-    LOSE;
-}
