+++ src/main/java/domain/WinState.java
@@ -1,0 +1,8 @@
+package domain;
+
+public enum WinState {
+
+    WIN,
+    DRAW,
+    LOSE;
+}
