+++ src/main/java/domain/cards/Card.java
@@ -1,0 +1,35 @@
+package domain.cards;
+
+import domain.cards.cardinfo.CardNumber;
+import domain.cards.cardinfo.CardShape;
+
+public class Card {
+
+    private final CardNumber cardNumber;
+    private final CardShape cardShape;
+
+    public Card(CardNumber cardNumber, CardShape cardShape) {
+        this.cardNumber = cardNumber;
+        this.cardShape = cardShape;
+    }
+
+    public boolean isAce() {
+        return cardNumber.equals(CardNumber.A);
+    }
+
+    public boolean isNotAce() {
+        return !cardNumber.equals(CardNumber.A);
+    }
+
+    public int score() {
+        return cardNumber.getScore();
+    }
+
+    public CardNumber getCardNumber() {
+        return cardNumber;
+    }
+
+    public CardShape getCardShape() {
+        return cardShape;
+    }
+}
