+++ src/main/java/domain/cards/gamercards/PlayerCards.java
@@ -1,0 +1,11 @@
+package domain.cards.gamercards;
+
+import domain.cards.Card;
+
+import java.util.List;
+
+public class PlayerCards extends GamerCards {
+    public PlayerCards(List<Card> cards) {
+        super(cards);
+    }
+}
