+++ src/main/java/domain/gamer/Player.java
@@ -1,0 +1,10 @@
+package domain.gamer;
+
+import domain.cards.gamercards.PlayerCards;
+
+public class Player extends Gamer {
+
+    public Player(String name, PlayerCards cards) {
+        super(name, cards);
+    }
+}
